// packages/frontend/e2e-smoke-tests/interfaces/page-interfaces.ts
export enum EPageUrl {
    daily = 'daily',
    monthly = 'monthly',
    terminal = 'terminal',
    dashboard = 'dashboard',
    balanceMonitor = 'balance-monitor',
    internalTransfers = 'internal-transfers',
    transferBlockingRules = 'transfer-blocking-rules',
    amountLimitsRules = 'amount-limits-rules',
    autoTransferRules = 'auto-transfer-rules',
}